requirements.txt
Flask
tensorflow
numpy
Pillow
flask-cors
